Slight modifications to locked_table to prepare for more operations
Replace buckets_ reference with a map_ reference, so we can call table
operations on it. Also get rid of check_table() call before every operation,
since we don't want to be doing such a check before every operation, and it's
not unreasonable to expect users to keep track of the validity of their locked
table.