Fix some allocator issues in the table
Most importantly, we force explicit moving of buckets during resize. We can't do
it in a move constructor, since we want to use the allocator provided to do the
construction. But we have to explicitly call move on each of the key-value pairs
that have been constructed in aligned_storage, so explicit reallocate, move, and
swapping of vectors is necessary.

Other small changes, like using std::allocator_traits everywhere, which will
fill holes in a custom the allocator definitions with sensible defaults.