Avoid generating a unique value for each insert operation
The content of the values doesn't actually matter, so when we pre-generated the
values, we were generating a bunch of copies of the same data, and then
re-copying that into the table.

Instead, we create a thread-local instance of the value and insert that
repeatedly. Technically, in the case of a pointer value, the pointer memory will
be deallocated when the thread finishes. But since the table shouldn't be
reading into the raw pointer memory, it shouldn't be a problem.